refactor: move patches to namespace
