+++ BlasII.Randomizer/Extensions/InfoExtensions.cs
